+++ src/usecase/recruitment/join_recruitment.go
@@ -1,0 +1,1 @@
+package recruitment
